fix: fix issues with data dependency in token prediction

+++ core/llm_token_ratelimit/rule_matcher.go
@@ -28,7 +28,7 @@ type MatchedRule struct {
 	CountStrategy CountStrategy
 	// PETA
 	Encoding       TokenEncoding
-	EstimatedToken int
+	EstimatedToken int64
 }
 
 type MatchedRuleCollector interface {
